rename const folder as constants and fix chat button
