+++ components/api/pageHeader.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import Link from 'next/link';
 
-import constants from '../../const';
+import constants from '../../constants';
 
 const Title = ({ title, logo }) => (
   <>
