+++ components/chat/chatStyles.scss
@@ -1,5 +1,5 @@
-@import '../../const/colors.scss';
-@import '../../const/layout.scss';
+@import '../../constants/colors.scss';
+@import '../../constants/layout.scss';
 
 #chat-container {
   z-index: 999;
