+++ components/home/useCaseSectionStyles.scss
@@ -1,4 +1,4 @@
-@import '../../const/colors.scss';
+@import '../../constants/colors.scss';
 
 section#use-cases {
   > .use-case {
