+++ components/searchApis/filterHeaderStyles.scss
@@ -1,4 +1,4 @@
-@import '../../const/colors.scss';
+@import '../../constants/colors.scss';
 
 section#search-api-filters {
   padding-bottom: 20px;
