+++ components/searchApis/resultsStyles.scss
@@ -1,4 +1,4 @@
-@import '../../const/colors.scss';
+@import '../../constants/colors.scss';
 
 section#api-results {
   padding: 35px 0 60px;
