+++ layouts/layout.scss
@@ -1,4 +1,4 @@
-@import '../const/colors.scss';
+@import '../constants/colors.scss';
 
 .layout-center {
   display: flex;
