+++ uiComponents/button/buttonStyles.scss
@@ -1,4 +1,4 @@
-@import '../../const/colors.scss';
+@import '../../constants/colors.scss';
 
 a.button-link,
 button.button-link {
