+++ uiComponents/dropdown/styles.scss
@@ -1,4 +1,4 @@
-@import '../../const/colors.scss';
+@import '../../constants/colors.scss';
 
 .dropdown-wrapper {
   label {
