+++ uiComponents/searchBar/index.js
@@ -2,7 +2,7 @@ import React, { useRef, useState } from 'react';
 import { debounce } from 'lodash';
 
 import { Search } from 'react-feather';
-import constants from '../../const';
+import constants from '../../constants';
 import './searchBarStyles.scss';
 
 const SearchBar = ({
