+++ uiComponents/searchBar/searchBarStyles.scss
@@ -1,4 +1,4 @@
-@import '../../const/colors.scss';
+@import '../../constants/colors.scss';
 
 #search-wrapper {
   max-width: 450px;
