+++ uiComponents/toggle/styles.scss
@@ -1,4 +1,4 @@
-@import '../../const/colors';
+@import '../../constants/colors';
 
 .toggle-wrapper {
   cursor: pointer;
